This compressed version not working
Hi,
I download this plugin from and First time I use this file with the latest version of jQuery. But I saw it not working. Then I use the non compressed version(jquery.parallax.js) this work fine with latest version of jQuery. Then I compressed the jquery.parallax.js file and I use the code which compressed by me. That time it work. It means that here it a problem in the current compressed file(jquery.parallax.min.js). 

So I request you to change the file. I change the code with the code which compressed by me.

Thank you.